moved requests to routes file
