+++ Chat/routes.js
@@ -1,68 +1,7 @@
 const express = require("express");
-const cors = require("cors");
-const jwt = require("jsonwebtoken");
-const mysql = require("mysql2/promise");
-const cookieParser = require("cookie-parser");
-require("dotenv").config();
+const router = express.Router()
 
-const port = 3 + +process.env.SERVER_PORT;
-
-const app = express();
-
-const pool = mysql.createPool({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-});
-
-app.use(async function (req, res, next) {
-  try {
-    req.db = await pool.getConnection();
-    req.db.connection.config.namedPlaceholders = true;
-
-    await req.db.query(`SET SESSION sql_mode = "TRADITIONAL"`);
-    await req.db.query(`SET time_zone = '-8:00'`);
-
-    await next();
-
-    req.db.release();
-  } catch (err) {
-    console.log(err);
-
-    if (req.db) req.db.release();
-    throw err;
-  }
-});
-
-app.use(express.json());
-app.use(cookieParser());
-app.use(
-  cors({
-    origin: `http://localhost:${process.env.CLIENT_PORT}`,
-    credentials: true,
-  })
-);
-
-function authenticateToken(req, res, next) {
-  const token = req.cookies.token;
-  if (token == null) {
-    return res.sendStatus(401);
-  }
-
-  jwt.verify(token, process.env.JWT_KEY, (err, user) => {
-    if (err) {
-      console.log(err);
-      return res.sendStatus(403);
-    }
-    req.user = user;
-    next();
-  });
-}
-
-app.use(authenticateToken);
-
-app.post("/loadTeamChat", async (req, res) => {
+router.post("/load", async (req, res) => {
   try {
     const { teamUID, teamName } = req.body;
     const userId = await findUserId(req);
@@ -92,7 +31,7 @@ app.post("/loadTeamChat", async (req, res) => {
   }
 });
 
-app.post("/createTeamChat", async (req, res) => {
+router.post("/create", async (req, res) => {
   try {
     const { message, teamUID, teamName } = req.body;
     const uid = Array.from(Array(254), () =>
@@ -127,7 +66,7 @@ app.post("/createTeamChat", async (req, res) => {
   }
 });
 
-app.post("/editTeamChat", async (req, res) => {
+router.post("/edit", async (req, res) => {
   try {
     const { chatUID, teamUID, teamName, message } = req.body;
     const userId = await findUserId(req);
@@ -156,7 +95,7 @@ app.post("/editTeamChat", async (req, res) => {
   }
 });
 
-app.delete("/deleteTeamChat", async (req, res) => {
+router.delete("/delete", async (req, res) => {
   try {
     const { chatUID, teamUID, teamName } = req.body;
     const userId = await findUserId(req);
@@ -191,11 +130,6 @@ app.delete("/deleteTeamChat", async (req, res) => {
   }
 });
 
-//Listener
-app.listen(port, () =>
-  console.log(`Chat server listening on http://localhost:${port}`)
-);
-
 //Functions
 //retrieves users id from the stored cookie
 async function findUserId(req) {
@@ -229,3 +163,5 @@ async function findJoinedTeamId(userId, teamUID, teamName, req) {
 
   return team || {};
 }
+
+module.exports = router
